requirements.txt
# Morgan Stanley Global Markets Analytics - Core Dependencies
# Financial Data Analysis & Risk Management

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Financial Analytics
yfinance>=0.2.0
pandas-datareader>=0.10.0
quandl>=3.7.0
pyfolio>=0.9.2

# Risk Management & Portfolio Analytics
empyrical>=0.5.5
pyfolio>=0.9.2
cvxpy>=1.3.0
scikit-learn>=1.3.0

# Database & SQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.0.0

# Time Series Analysis
statsmodels>=0.14.0
arch>=6.2.0

# Compliance & Audit
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
